+++ harfbuzz/src/hb-ot-layout.cc
@@ -40,7 +40,6 @@
 #include "hb-ot-layout-gdef-table.hh"
 #include "hb-ot-layout-gsub-table.hh"
 #include "hb-ot-layout-gpos-table.hh"
-#include "hb-ot-os2-table.hh"
 
 #include "hb-aat-layout-morx-table.hh"