+++ CHANGELOG.md
@@ -6,16 +6,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Ported
+- Tables: `OS/2`
 - `hb_script_t`
 - `hb_feature_t`
 - `hb_variation_t`
 - `hb_font_t`
+- `hb-ot-metrics`
 
 ### Changed
 - Rename `Font::from_data` into `Font::from_slice`.
 - Font is parsed via `ttf-parser` first.
   And if the parsing fails, the `Font` will not be created.
-  `harfbuzz` allows malfomed fonts.
+  `harfbuzz` allows malformed fonts.
 
 ### Removed
 - `hb_font_funcs_t`. Only the embedded TrueType implementation is used.